refactor: consolidate shared utilities and remove obsolete helpers.py
- Move get_temperature_sensor_entities function from helpers.py to global_config_entity.py
- Update all import references across the codebase to use the new location
- Remove obsolete helpers.py file after consolidating shared functionality
- Update test patches to reference the correct module paths
- All tests passing, code duplication reduced by ~37%

This completes the comprehensive refactoring of 6 platform files to use shared base classes and consolidated utilities.
