+++ tests/config_flow/test_string_conversion_comprehensive.py
@@ -41,7 +41,7 @@ class TestSecondSaveBugFix:
                 return_value=config_entries.SOURCE_USER,
             ),
             patch(
-                "custom_components.solar_window_system.helpers.get_temperature_sensor_entities",
+                "custom_components.solar_window_system.global_config_entity.get_temperature_sensor_entities",
                 return_value=[],
             ),
         ):
@@ -80,7 +80,7 @@ class TestSecondSaveBugFix:
                 return_value=AsyncMock(data=stored_data),
             ),
             patch(
-                "custom_components.solar_window_system.helpers.get_temperature_sensor_entities",
+                "custom_components.solar_window_system.global_config_entity.get_temperature_sensor_entities",
                 return_value=[],
             ),
         ):
@@ -119,7 +119,7 @@ class TestSecondSaveBugFix:
                 return_value=config_entries.SOURCE_USER,
             ),
             patch(
-                "custom_components.solar_window_system.helpers.get_temperature_sensor_entities",
+                "custom_components.solar_window_system.global_config_entity.get_temperature_sensor_entities",
                 return_value=[],
             ),
             patch.object(
@@ -167,7 +167,7 @@ class TestSecondSaveBugFix:
                 return_value=AsyncMock(data=stored_data),
             ),
             patch(
-                "custom_components.solar_window_system.helpers.get_temperature_sensor_entities",
+                "custom_components.solar_window_system.global_config_entity.get_temperature_sensor_entities",
                 return_value=[],
             ),
             patch.object(
